fix: increased the maximum amount of the price of products

+++ models/receipt.js
@@ -15,7 +15,7 @@ export const Receipt = sequelize.define("receipt", {
             defaultValue: new Date().toISOString().slice(0, 19).replace('T', ' ')
         },
         total: {
-            type: Sequelize.DOUBLE(7, 2),
+            type: Sequelize.DOUBLE(14, 2),
             defaultValue: null
         }
     },
